lib: fix atomic api and add spinlock api

+++ include/aco/atomic.h
@@ -1,7 +1,7 @@
 #ifndef __ACO_ATOMIC_H
 #define __ACO_ATMOIC_H
 
-typedef unsigned int atomic_t;
+typedef unsigned short atomic_t;
 
 // if we use gcc , use gcc built-in atomic operation
 // else take it to specific arch to implement it
